Rebalance Will roll data
This change should make most panic rolls a bit easier to succeed on. I
have also updated the panic types so each Will roll type results in a
specific panic type.

Lastly, I updated Horror so that it only reduces Will by half the max
Will of the target, and the chance to panic has been reduced so that it
could be resisted if you're lucky. The aim is to try to differentiate
this from a normal panic-causing ability (like Mindspin), while making
it a little less OP.