Fix RespecSoldier command for AP cost escalation
Invoking any of the soldier respec commands reset all the ability
selections, but it did not clear the unit value that tracked how many
class abilities at a given rank had been picked. This meant that if you
had previously incurred an AP cost multiplier because you'd picked a
class perk at a rank that already had one picked, then the multiplier
was remembered and the next time after respec you tried the same thing,
the multiplier was higher than it should be.

This was fixed by clearing the 'LWOTC_AbilityCostModifier' unit value
from the soldier when respeccing.