Fix Templar's Void Conduit ability (#951)
For an explanation of the fix, see the `FixVoidConduit()` function in
`X2LWModTemplate_TemplarAbilities`. It basically ensures that the
targeted unit loses the appropriate number of action points at the start
of its turn.

I have also added a flyover that shows how many actions the target
has left after Void Conduit ends. Otherwise the animation for Void
Conduit ends right at the beginning of the enemy turn and it looks
like they can take actions. So this flyover just acts as a "yes, the
effect has ended, but actions were removed from the target, so don't
worry".