Fix Chosen details button on tactical HUD (#249)
The officer pack was overriding `UITacticalHUD_MouseControls` and only
had the original XCOM 2 implementation. Hence the code to display the
Chosen button was missing.

The override now calls the super class function so that the Chosen
button is added properly. It also positions the Command Range button
differently based on whether the Chosen are enabled or not (it needs to
leave enough space for the Chosen button).

Fixes #249.