Fix hacks extending Network Tower stun (#583)
Shutting down a robotic enemy successfully after hacking the Network
Tower objective was causing the Network Tower hack reward ability
('NetworkTowerStun_LW') to trigger again. This is because once the
ability is applied to the target (the hacker in this case), it triggers
for every hack.

To fix this, I have given the NetworkTowerStun_LW ability a single
charge so that it can only activate once in a mission.

Fixes #583.