Fix M3 enemies appearing on Rendezvous
As part of the second go at fixing #140 I reintroduced the call to
`UpdateMissionSchedules()` in the `UIMission` screens. However, this was
a bad idea as it updates the mission schedules using the vanilla force
and alert levels. So Rendezvous missions were getting current FL
schedules if the Shadow Chamber was built, rather than the reduced FL
they were supposed to.

I've reverted that part of the fix so that `UpdateMissionSchedules()` is
bypassed again.