Fix Chosen info icon w/ Smart Overwatch All
The Chosen info icon was always displaying in tactical, even if the
Chosen were disabled. This was because Smart Overwatch All adds a
command ability, and we were setting the number of active controls for
UITacticalHUD_MouseControls based on the number of command abilities.

However, the extra command ability was being display using the Chosen
info icon. A bit weird, but command abilities aren't well implemented.
Anyway, the base game only adds one to the number of active controls, no
matter the number of command abilities (unless there are none, in which
case it doesn't add to the number). So I've fixed the Officer Pack code
to do the same.