Fix controller selecting infiltration options (#961)
-----------------------------------------

Modifies UIUtilities_LW --> BuildMissionInfoPanel().

When you click on a mission which is being infiltrated the infiltration screen,
UIMission_LWLaunchDelayedMission, pops up. The function BuildMissionInfoPanel()
creates the special mission information panel seen near the top of this screen.
The problem is that within this function, a UIPanel and UIX2PanelHeader are
created; however, they are never removed from the navigation system and end up
causing problems with controller and arrow key navigation. Consequently,
bIsNavigable has been set to false for both of them so they are no longer
added to the navigation system.

-----------------------------------------

Modifies UIStrategyMapItem_Mission_LW --> OnMouseEvent() and OnUnrealCommand().

Originally, when a controller user clicked on the A button while a map item
mission was highlighted, it would open a normal mission screen even if the
mission was currently undergoing infiltration. As a result, OnUnrealCommand()
had to be extended so as to call the same code as found in OnMouseEvent() -->
FXS_L_MOUSE_UP. This code calls OpenInfiltrationMissionScreen(), which pushes
an infiltration screen onto the screen stack if infiltration is occurring.

-----------------------------------------

Modifies UIMission_LWLaunchDelayedMission.

The main changes occur in BindLibraryItem(), RefreshNavigation(), and
OnUnrealCommand().

1.] BindLibraryItem() has been overridden so that all of the UIButtons
("BOOST INFILTRATION", "VIEW SQUAD", "ABORT", "LAUNCH MISSION", and "WAIT") are
created in the same manner whether the user is using a controller or mouse &
keyboard; the main change is that they all have the style eUIButtonStyle_NONE.
This guarantees that each button displays itself as a normal button rather than
a hotlink; since there are up to 5 selectable buttons on this screen,
controller hotlinks simply won't work.

2.] RefreshNavigation() has been overridden since the UIMission version is a
bit of a mess, and to make it compatible with a controller and the arrow keys.
It now clears the navigation system, adds any visible buttons to the navigation
system, and then selects an initial button depending upon which buttons are
visible.

3.] OnUnrealCommand() has been overridden so that the A button and
Enter/Spacebar key click on whichever button happens to be selected.

There has also been a bit of cleanup and a fix for a problem in which the
IgnoreButton could spawn several times on top of itself.