Remove problematic Geoscape help icons when using a controller
-------------------------

Modifies UIStrategyMapItem_Region_LW --> InitMapItem()

When using a controller, a (contacted) region map item can show 3 possible
help icons.

1.] MapItem_Region --> consoleHint is the help icon which appears to the left
of the radio tower button.
2.] StrategyScanButton --> scanHint is the help icon which appears to the right
of the scan button, when visible.
3.] StrategyScanButton --> consoleHint appears to the left of the haven
information button when no scan button is visible.

The problem with these help icons is that sometimes they display the wrong
icon, and sometimes they overlap other UI elements. Hiding them is impossible
due to the underlying flash code; consequently, the easiest way to deal with
them is to remove via setting them to null.

-------------------------

Modifies UIStrategyMapItem_Mission_LW --> InitMapItem()

When using a controller, an infiltrating mission displays a help icon,
StrategyScanButton --> consoleHint, to the left of the haven information
button. The problem is that it overlaps the icon representing the infiltrating
squad. Consequently, just like above, the best way to deal with it, is to null
it.