Modify Reveal Unit effect to work around crash
In some cases, loading a tactical save with a Reaper on it that has
Tracking will crash because the game state for Tracking's Reveal Unit
visualisation action has a context that's `none`. I don't yet know why
or how that happens, but this change works around the problem by
checking in the visualisation action for a `none` context and skipping
the visualisation in that case.