Make AI slightly smarter against the Lost
Thanks to Nilch for this suggestion to heavily reduce the targeting
score for the Lost in the AI's behaviour trees. This should make ADVENT
target XCOM more often than before when the Lost are around. This makes
the Lost less of a "safe" sit rep for XCOM as the Lost can no longer be
used as a reliable distraction for the more dangerous ADVENT and alien
units.