Rebalance the SMG
The SMG now gains +1 ammo so that it can be used with some of the
cross-class abilities that require 3 ammo. To balance it, the weapon
loses 1 from its Mobility bonus, but has better close range Aim bonuses.

This also updates the Bullpup and Templar pistol range tables to match
the new SMG one.