Reposition scrollbar on promotion screen
It was running down the middle of the MSGT abilities, so I shifted it
right a bit. I have also added a `ScrollbarPositionX` config variable
that can be used to override the default position, in case there are
some differences between LWOTC setups that means the promotion screen is
wider/narrower or appears in a different location.