Fix infiltrators from DEs appearing on missions (#1140)
`CacheSelectMissionData()` populates the encounters and checks whether a
tactical tag is active before adding conditional encounters like the
ones used for the Infiltrator dark events (Faceless and Chryssalids).
The issue was that the function was checking the XCOM HQ tactical tags,
not the ones in the mission, and the HQ tags don't include the sit rep
tags when `CacheSelectMissionData()` is called.

This fix simply copies the mission tactical tags to XCOM HQ immediately
before `CacheSelectMissionData()` is called.

Fixes #1140.