Fix dark event sit reps appearing on invalid missions
While we were checking to ensure that dark event sit reps satisfied the
LWOTC-specific configuration with regards to the validity of sit reps
for a given mission type, we were ignoring the base game configuration,
like `X2SitRepTemplate.ValidMissionTypes`.

We now use `X2SitRepTemplate.MeetsRequirements()` as the authority on
whether a sit rep is valid for a given mission. The LWOTC-specific
checks are incorporated by a community-highlander event hook in
`MeetsRequirements()`.