Nerf Vigilance and Apotheosis slightly
Vigilance will now only activate if the pod that's revealed has at least
3 units in it. The minimum number of units in the pod is configurable.

Apotheosis now factors in Reaper if the latter ability is active. That's
because previously Apotheosis could be activated before Reaper,
resulting in the doubling of damage happening before the Reaper damage
reduction kicked in. This meant that Templars could sustain high damage
on long Reaper chains.

This fix (hack) checks whether Reaper is active and is applied after
Apotheosis. In that case, the Reaper damage reduction is applied before
Apotheosis doubles the damage. So the Reaper reduction is doubled too.