Fix Infiltrator officer ability
The code for calculating squad covertness was using the
`XComGameState_Unit.HasSoldierAbility()` function to determine whether a
unit had Infiltrator or not. Unfortunately, that function doesn't take
into account abilities that mods add to units via
`XCGS_Unit.GetEarnedSoldierAbilities()`, which in this case includes the
officer abilities.

The fix is a bit of a hack, but now the squad covertness calculation
uses a custom version of `HasSoldierAbility()` copied from Long War 2
that iterates directly over the earned soldier abilities.