Allow M3 Gunner to Traverse Fire (#1152)
TryTraverseFire was using IsAbilityAvailable-TraverseFire which was not working due to it being a passive skill. Replaced with TemplateNameIs-AdvGunnerM3.
IsAbilityAvailable-TraverseFire was deleted as it was non-functional condition.
Added TryHighPriorityShot on LWAdventGunnerRedFirstActionSelector so AI can guarantee shoot-shoot actions against vulnerable targets.
Added LWAdventGunnerRedPostTraverseFire because LWAdventGunnerRedLastAction did not recognize action refunded by Traverse Fire and returned false.