Fix Revival Protocol and Solace restoring action points
Revival Protocol was failing to restore the action points of stunned
units if the target had been stunned for at least a whole turn before
being revived. This was because the condition for whether the target was
stunned or not wasn't quite correct.

Solace was allowing a disoriented unit to use all its action, be
cleanse, and then have its action points restored. Now Solace will not
restore action points to disoriented units.