Attempt to make Alien Rulers appear more often
The previous attempt to do this using higher weights for the Alien
Rulers didn't seem to have any effect at all, possibly because the card
manager does not work the way I hoped.

This change provides a separate mechanism for adding sit reps to
missions that can be used for sit reps that would otherwise be very rare
due to the restrictions on when they can apply. These "special" sit reps
are rolled separately *before* the normal sit rep allocation kicks in
and allows for separate chances for each such special sit rep.

I have also added a couple of event hooks that allow mods to influence
the sit reps on missions.