Fix bondmate ability to properly clear stuns
The bondmate ability based on Solace was clearing the stun effect, but
not the stunned action points for the target unit. So the unit would
have the effect removed and would stand up, but would have no action
points to do anything. This would last as long as the original stun was
due to last.

The ability now applies `X2Effect_StunRecover` to the target, which
restores the action points.