Rework the covert action Ambush mission
The mission now consists of the commander sending a squad of four
soldiers to rescue the covert action operatives. The mission, and by
extension the covert action, will fail if not a single operative evacs
out. If the covert action fails, the rewards will not be given to the
player.

I've updated the mission screen/blade so that the player can back out of
it in order to find suitable soldiers, for example by pulling them out
of training, haven duty or infiltrations. The mission is available for
12 hours before expiring. If it expires, the covert action fails and the
operatives are captured.