Fix Chosen not appearing on their on Avenger assault
We were removing the Chosen tag in `OnPreMission()` but never adding it
back because we weren't adding the LWOTC-specific Chosen tag to the
mission. Rather than mess around with the base-game code that launches
the mission, we now just let the base game do its thing for select
missions, including the Chosen Avenger Defense.