Allow rescuing multiple captured soldiers
Jailbreaks can now have multiple captured soldiers as rewards on a
single mission, making it far less troublesome to recover soldiers when
the Chosen have been kidnapping a whole bunch of them.

A lot of this code is designed to ensure that each captured soldier can
only be the reward for a single mission or covert action. Neither
vanilla not LW2 handled this properly.