Fix Chosen appearing on Alien Ruler missions
`OnPreMission()` was bypassing the removal of the vanilla-added Chosen
tags if an Alien Ruler was on a mission, resulting in the Chosen
potentially appearing on the same mission as the Ruler.

Now the Chosen tags will be removed even if a Ruler is already on the
mission, but they won't be re-added if that's the case.