Improve WOTC overwatch with cover defense reduction
If Revert Overwatch Rules mod is not being used, this change applies a
reduction to the Defense bonus from cover, making the WOTC overwatch
handling a lot better. Overwatch shots against units in cover are no
longer seriously disadvantaged.

This has been achieved by applying an effect to all unit types that can
take cover. The effect checks whether the unit it's affecting is in
cover relative to the attacker.

Revert Overwatch Rules remains a "required" mod, but players that are
happy to experiment can deactivate that mod and immediately get to play
with this new behaviour.