Fix Soul Steal appearing twice in training screen
A vanilla bug in `UIChoosePsiAbility.GetAbilities()` was causing Soul
Steal to occasionally appear twice as a training option because it has
Soulfire as a prerequisite. Removing the prerequisite fixes the problem
for now, but it should really be properly fixed in the highlander.

Note that removing Soulfire as the prerequisite *only* works because
every psi operative has Soulfire as a base ability.