Fix duration of Maimed status effect
One turn is not sufficient if the effect ticks at the start of the
player turn, since then the effect is removed just before the target
gets to take their actions.

This change ups the duration to 2 turns, just like with Disoriented and
Burning, and adds the duration to config.