Fix timer on Defend_LW to use config correctly
Defend_LW now pulls the initial reinforcement timer value via
`SeqAct_GetMissionTimer`, which is the authority in LWOTC.

I also fixed the mission to work properly with the
`EngageMissionTimerAtStart` config setting. This prevents the mission
from decrementing the timer on the first turn. The initial turn timer
value has been reduced by 1 to compensate for this change.