Fix stacking of Maimed status effect
The mobility multiplier was being added in `OnEffectAdded`, which meant
is was being applied every time the Maimed effect was. This was causing
the multipliers to stack.

The multiplier is now set when the effect is created, but now there are
two effects: one for the Chosen (that reduces mobility by 50%) and one
for everyone else (that reduces mobility to zero).