Add compatibility with extra inventory slots
This will add compatibility for Spark Arsenal's Auxiliary Slot, Spark Grenade Slot, and BIT Heavy Weapon Slot, as well as Psionics Overhaul V3's Psi Amp Slot, so that none of those equipment slots can be messed with on a soldier that's infiltrating. I also removed the duplicate GrenadePocket line because it doesn't really need to be there.