Disable shadow chamber mission info pre-infiltration
The Shadow Chamber panel for infiltration missions that don't have
squads attached to them yet doesn't necessarily show the enemies that
the player will encounter when it's launched. This can be confusing,
since the player may prepare a squad based on the enemies shown.

This change simply hides the panel for mission screens/blades that
aren't yet being infiltrated. It's shown for missions with infiltrating
squads because it correctly shows the enemies you'll encounter if you
press the launch button right then.