Fix covert action risk chances when negating risks
Negating a risk was causing the ambushed chance to be displayed as the
chance of the risk that was negated. For example, if Captured had a
chance of 10% and Ambush had a chance of 21%, then negating Captured
would then have Ambush displayed as 10%.

This was because the listener that overrides the labels didn't take into
account that the array of labels passed to it may not be the same length
as the array of risks. The code no longer makes that assumption.