Unifying cooldown description format
Replaces many instances of plain text and ablity-specific variables for abilities cooldown descriptions.
Introduces <Ability:SelfCooldown_LW> tag which does not subtract 1 from original cooldown value, unlike the original SelfCooldown.
Standartizes the manner in which cooldown is described both in the  LocPromotionPopupText and in certain instances of LocHelpText.
Fixes couple of typos.