Fixing some misinformation in descriptions
Changes were made to bring abilities' descriptions in line with how they actually work. In some cases it is clear that the intended way for an ability to work was "as described" but for some reason or another the implementation does not work that way. If implementation of these abilities were to be fixed later, then some descriptions might need to be reverted back.

The list of changed descriptions and what was wrong with them:
Point Blank and Both Barrels - do end the turn.
Holotargeting - lasts for 1 enemy turn as well and does end the turn by default.
Whirlwind - does in fact work with Reaper active. Despite having check against it in the code, it somehow fails.. Might want to be looked at.
Focus - had a lot of pointless lines for unused focus levels.
Manual Override - since it is now affect allies cooldowns, its original usage recommendation is outdated.
Close Combat Specialist - does work on your turn for now. Might want to be looked at.
Light 'Em Up - it is not technically applies to only first shot. Marauder wording is more precise.
Suppression and Area Suppresion - multiple effects do not stack.
Boosted Cores - does not apply to grenade effects.
Concealment - cannot in fact be used at "any" location, namely if the unit is flanked already.