Tweak ambush mission settings
Reinforcements should be delayed by one extra turn now and I have made
the pod that starts close to the covert operatives a bit easier for both
easy and difficult ambushes (but not moderate ones). For difficult ones,
I moved one unit from that pod to the one that spawns near evac.

I have also shifted the two pods furthest from the covert operatives to
slightly further away in the hopes that takes some of the pressure off
them.