Make Resistance/Outpost Screen controls a bit more consistent.
When viewing the strategy map, the X button brings up the Resistance Management screen. Consequently, I think it is more consistent if, while on the Resistance Management screen, the X button brings up the selected haven rather than the A button. Furthermore, it feels better if pressing the X button on a Haven screen modifies the haven adviser, rather than through right stick click.

------------------------------

Modifies : UIResistanceManagement_LW.uc

On the Resistance Management screen the A button sends the Avenger to the selected haven; formerly, this was attached to right stick click.

------------------------------

Modifies : UIResistanceManagement_ListItem.uc

On the Resistance Management screen the X button opens the corresponding haven screen; formerly, this was attached to the A button.

------------------------------

Modifies : UIOutpostManagement.uc

When viewing a particular haven, the X button now [1] Opens the haven adviser selection screen, if no haven adviser currently exists. [2] Removes the haven adviser, if one currently exists. Formerly, this was attached to right stick click.