Add 'Large Information Panel' option on mission screen
The LW2 mission information panel, displayed when clicking on a mission on the strategy layer, has text which is nearly unreadable for people not sitting right in front of a monitor. This provides an alternative information panel which is a bit larger and, more importantly, has larger text.

------------------------------

Modifies : UIVerticalScrollingText2.SetText and UIVerticalScrollingText2.SetHTMLText

UIVerticalScrollingText2 is a LW created text container which fixes UITextContainer's inability to display scrolled text; when using a UITextContainer, text which starts outside of the text container's visible area would remain invisible even when scrolling into view.

The one problem with UIVerticalScrollingText2 was that, upon scroll completion, a new 'scroll cycle' would commence with the text being lower than it should. This was because, within SetText and SetHTMLText, textPosOffset was used to modify the text position even if no background box, bg, existed. On the other hand, within RealizeTextSize, textPosOffset was set to 0 if no background box existed.

Here is what happened in more detail :
1.] UIVerticalScrollingText2.SetHTMLText was called to modify the UIText, "text".
2.] Within UIVerticalScrollingText2.SetHTMLText, "text"'s position was offset by textPosOffset and then UIText.SetHtmlText was called.
3.] Within UIText.SetHtmlText, Flash's TextControl.setText was called; its first parameter was the text to display, and the second paramter was the callback function UIVerticalScrollingText2.RealizeTextSize.
4.] Once the text had been realized in Flash, UIVerticalScrollingText2.RealizeTextSize was called.
5.] Since bAutoScroll was true, UIPanel.AnimateScroll was called.
6.] UIPanel.AnimateScroll would call Flash's Panel.animateScroll which would then call Flash's Utilities.AnimateScroll
7.] Utilities.AnimateScroll would end up calling "object.UTIL_initialObjectY = object._y"; basically it would store the UIText's current Y location as the Y value it should be reset to upon 'scroll completion'.

The crux of the problem is that in step 2 the UIText was modified, along the Y axis, by textPosOffset and this offset was saved in step 7. HOWEVER, the original Y location of the UIText is set at the bottom of UIVerticalScrollingText2.RealizeTextSize with an offset of 0.

------------------------------

Modifies : XComLW_UI.ini

Adds 2 new config variables : USE_LARGE_INFO_PANEL, and CENTER_TEXT_IN_LIF; these default to false. USE_LARGE_INFO_PANEL displays the large information panel rather than the normal LW2 information panel. CENTER_TEXT_IN_LIF center-aligns the information panel text if and only if the large information panel is being used.

------------------------------

Modifies : UIUtilities_LW.uc
Cleans up UIUtilities_LW.BuildMissionInfoPanel and adds UIUtilities_LW.GetMissionInfoPanelText

The LW2 variant of BuildMissionInfoPanel was quite difficult to follow; therefore, BuildMissionInfoPanel was cleaned up. Now, BuildMissionInfoPanel focuses on UI creation while GetMissionInfoPanelText focuses on string gathering for these UI elements.

Logic from the original UIUtilities_LW.BuildMissionInfoPanel was carefully transferred over so that the "normal" information panel appears exactly like it did within LW2. The large information panel retains the location of the infiltration information and mission expiration information; however, it adds a dividing line and places the "other" mission information below this dividing line. Additionally, as mentioned above, it makes use of larger text.