Fix Quick Zap ability
Quick Zap was not configured with any abilities that it applied to, so
it did nothing. I renamed the relevant `X2Condition` so that it wouldn't
conflict with the original, pointed it to XComLW_SoldiersSkills.ini, and
added the appropriate abilities to that config.

I also updated the ability icon code so that it displays the arc thrower
icon in green when Quick Zap is active.